[Ashish] Add:UC2 Solved

+++ MaximumNumber/TestMaximum.cs
@@ -8,7 +8,7 @@ namespace MaximumNumber
 {
     public class TestMaximum
     {
-        public int MaximumNumber(int firstnumber ,int secondnumber, int thirdnumber)
+        public double MaximumNumber(double firstnumber ,double secondnumber, double thirdnumber)
         {
             if (firstnumber.CompareTo(secondnumber) > 0 && firstnumber.CompareTo(thirdnumber) > 0) return firstnumber;
             if (secondnumber.CompareTo(firstnumber) > 0 && secondnumber.CompareTo(thirdnumber) > 0) return secondnumber;
